♻️ refactor: correctr mispelling in AcquireAvatarUseCase and AcquireRocketUseCase
