+++ packages/core/src/profile/use-cases/AcquireAvatarUseCase.ts
@@ -3,34 +3,34 @@ import { Integer } from '#global/domain/structures/Integer'
 import { Id } from '#global/domain/structures/Id'
 import type { UsersRepository } from '../interfaces'
 import { UserNotFoundError } from '../errors'
-import { RocketAggregate } from '../domain/aggregates'
+import { AvatarAggregate } from '../domain/aggregates'
 import type { UserDto } from '../domain/entities/dtos'
 
 type Request = {
   userId: string
-  rocketId: string
-  rocketName?: string
-  rocketImage?: string
-  rocketPrice: number
+  avatarId: string
+  avatarName?: string
+  avatarImage?: string
+  avatarPrice: number
 }
 
 type Response = Promise<UserDto>
 
-export class AquireRocketUseCase implements UseCase<Request, Response> {
+export class AcquireAvatarUseCase implements UseCase<Request, Response> {
   constructor(private readonly repository: UsersRepository) {}
 
-  async execute({ userId, rocketId, rocketName, rocketImage, rocketPrice }: Request) {
+  async execute({ userId, avatarId, avatarName, avatarImage, avatarPrice }: Request) {
     const user = await this.repository.findById(Id.create(userId))
     if (!user) throw new UserNotFoundError()
 
     let entity = undefined
-    if (rocketName && rocketImage) {
-      entity = { name: rocketName, image: rocketImage }
+    if (avatarName && avatarImage) {
+      entity = { name: avatarName, image: avatarImage }
     }
-    const rocket = RocketAggregate.create({ id: rocketId, entity })
+    const avatar = AvatarAggregate.create({ id: avatarId, entity })
 
-    user.buyRocket(rocket, Integer.create(rocketPrice))
-    await this.repository.addAcquiredRocket(rocket.id, Id.create(userId))
+    user.buyAvatar(avatar, Integer.create(avatarPrice))
+    await this.repository.addAcquiredAvatar(avatar.id, Id.create(userId))
     return user.dto
   }
 }
