+++ packages/core/src/profile/use-cases/AcquireRocketUseCase.ts
@@ -3,34 +3,34 @@ import { Integer } from '#global/domain/structures/Integer'
 import { Id } from '#global/domain/structures/Id'
 import type { UsersRepository } from '../interfaces'
 import { UserNotFoundError } from '../errors'
-import { AvatarAggregate } from '../domain/aggregates'
+import { RocketAggregate } from '../domain/aggregates'
 import type { UserDto } from '../domain/entities/dtos'
 
 type Request = {
   userId: string
-  avatarId: string
-  avatarName?: string
-  avatarImage?: string
-  avatarPrice: number
+  rocketId: string
+  rocketName?: string
+  rocketImage?: string
+  rocketPrice: number
 }
 
 type Response = Promise<UserDto>
 
-export class AquireAvatarUseCase implements UseCase<Request, Response> {
+export class AcquireRocketUseCase implements UseCase<Request, Response> {
   constructor(private readonly repository: UsersRepository) {}
 
-  async execute({ userId, avatarId, avatarName, avatarImage, avatarPrice }: Request) {
+  async execute({ userId, rocketId, rocketName, rocketImage, rocketPrice }: Request) {
     const user = await this.repository.findById(Id.create(userId))
     if (!user) throw new UserNotFoundError()
 
     let entity = undefined
-    if (avatarName && avatarImage) {
-      entity = { name: avatarName, image: avatarImage }
+    if (rocketName && rocketImage) {
+      entity = { name: rocketName, image: rocketImage }
     }
-    const avatar = AvatarAggregate.create({ id: avatarId, entity })
+    const rocket = RocketAggregate.create({ id: rocketId, entity })
 
-    user.buyAvatar(avatar, Integer.create(avatarPrice))
-    await this.repository.addAcquiredAvatar(avatar.id, Id.create(userId))
+    user.buyRocket(rocket, Integer.create(rocketPrice))
+    await this.repository.addAcquiredRocket(rocket.id, Id.create(userId))
     return user.dto
   }
 }
