+++ packages/core/src/shop/use-cases/index.ts
@@ -1,2 +1,2 @@
-export { AquireDefaultShopItemsUseCase } from './AquireDefaultShopItemsUseCase'
+export { AcquireDefaultShopItemsUseCase } from './AcquireDefaultShopItemsUseCase'
 export { GetAcquirableShopItemsByDefaultUseCase } from './GetAcquirableShopItemsByDefaultUseCase'
